Factor out Object File handling from `swift-reflection-dump` to separate CMake target: StaticMirror
This target will then be used to have a broader set of tools for reading metadata from object files.
